FinanceAISearch/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
distlib==0.3.9
distro==1.9.0
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
identify==2.6.5
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
jwt==1.3.1
Mako==1.3.6
Markdown==3.7
MarkupSafe==3.0.2
multidict==6.1.0
nodeenv==1.9.1
numpy==2.2.1
oauthlib==3.2.2
openai==0.28.0
pandas==2.2.3
platformdirs==4.3.6
pre_commit==4.0.1
propcache==0.2.0
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
requests-oauthlib==1.3.1
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
ta==0.11.0
tqdm==4.67.0
tweepy==4.14.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.28.1
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.18.0
